refactor(#567): generalize drug list
