+++ app/lib/common/widgets/module.dart
@@ -1,5 +1,7 @@
 export 'app.dart';
 export 'context_menu.dart';
+export 'drug_list/builder.dart';
+export 'drug_list/cubit.dart';
 export 'headings.dart';
 export 'indicators.dart';
 export 'page_scaffold.dart';
