+++ app/lib/search/pages/search.dart
@@ -2,7 +2,6 @@ import 'package:flutter/cupertino.dart';
 
 import '../../../common/module.dart';
 import '../../common/pages/drug/widgets/tooltip_icon.dart';
-import 'cubit.dart';
 
 class SearchPage extends HookWidget {
   SearchPage({
@@ -27,26 +26,21 @@ class SearchPage extends HookWidget {
         child: BlocBuilder<DrugListCubit, DrugListState>(
             builder: (context, state) {
           return pageScaffold(
-              title: context.l10n.tab_drugs,
-              barBottom: Row(children: [
-                Expanded(
-                    child: CupertinoSearchTextField(
-                  controller: searchController,
-                  onChanged: (value) {
-                    context.read<DrugListCubit>().search(query: value);
-                  },
-                )),
-                SizedBox(width: 12),
-                TooltipIcon(context.l10n.search_page_tooltip_search),
-                buildFilter(context),
-              ]),
-              body: state.when(
-                initial: () => [Container()],
-                error: () => [errorIndicator(context.l10n.err_generic)],
-                loaded: (drugs, filter) =>
-                    _buildDrugsList(context, drugs, filter),
-                loading: () => [loadingIndicator()],
-              ));
+            title: context.l10n.tab_drugs,
+            barBottom: Row(children: [
+              Expanded(
+                  child: CupertinoSearchTextField(
+                controller: searchController,
+                onChanged: (value) {
+                  context.read<DrugListCubit>().search(query: value);
+                },
+              )),
+              SizedBox(width: 12),
+              TooltipIcon(context.l10n.search_page_tooltip_search),
+              buildFilter(context),
+            ]),
+            body: buildDrugList(context, state),
+          );
         }));
   }
 
@@ -71,84 +65,6 @@ class SearchPage extends HookWidget {
       ],
       child: Padding(
           padding: EdgeInsets.all(8), child: Icon(Icons.filter_list_rounded)),
-    );
-  }
-
-  List<Widget> _buildDrugsList(
-      BuildContext context, List<Drug> drugs, FilterState filter) {
-    final filteredDrugs = filter.filter(drugs);
-    if (filteredDrugs.isEmpty) {
-      return [errorIndicator(context.l10n.err_no_drugs)];
-    }
-    return [
-      SizedBox(height: 8),
-      ...filteredDrugs.map((drug) => Column(children: [
-            Padding(
-                padding: EdgeInsets.symmetric(horizontal: 8),
-                child: DrugCard(
-                    onTap: () => context.router
-                        .push(DrugRoute(drug: drug))
-                        .then((_) => context.read<DrugListCubit>().search()),
-                    drug: drug)),
-            SizedBox(height: 12)
-          ]))
-    ];
-  }
-}
-
-class DrugCard extends StatelessWidget {
-  const DrugCard({
-    required this.onTap,
-    required this.drug,
-  });
-
-  final VoidCallback onTap;
-  final Drug drug;
-
-  @override
-  Widget build(BuildContext context) {
-    final warningLevel = drug.userGuideline()?.annotations.warningLevel;
-
-    return RoundedCard(
-      onTap: onTap,
-      padding: EdgeInsets.all(8),
-      radius: 16,
-      color: warningLevel?.color ?? PharMeTheme.onSurfaceColor,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(children: [
-                  Icon(warningLevel?.icon ?? Icons.help_outline_rounded),
-                  SizedBox(width: 4),
-                  Text(
-                    drug.name.capitalize(),
-                    style: PharMeTheme.textTheme.titleMedium!
-                        .copyWith(fontWeight: FontWeight.bold),
-                  ),
-                ]),
-                SizedBox(height: 4),
-                if (drug.annotations.brandNames.isNotEmpty) ...[
-                  SizedBox(width: 4),
-                  Text(
-                    '(${drug.annotations.brandNames.join(', ')})',
-                    style: PharMeTheme.textTheme.titleMedium,
-                  ),
-                ],
-                SizedBox(height: 8),
-                Text(
-                  drug.annotations.drugclass,
-                  style: PharMeTheme.textTheme.titleSmall,
-                ),
-              ],
-            ),
-          ),
-          Icon(Icons.chevron_right_rounded),
-        ],
-      ),
     );
   }
 }
